Move deps inside tooltip
