--- src/ui/radix/tooltip/TooltipArrow.tsx
@@ -1,5 +1,0 @@
-import * as TooltipPrimitive from '@radix-ui/react-tooltip'
-
-export const TooltipArrow = ({ className }: { className?: string }) => (
-  <TooltipPrimitive.Arrow className={className} />
-)
