--- src/ui/radix/tooltip/TooltipTrigger.tsx
@@ -1,19 +1,0 @@
-import * as TooltipPrimitive from '@radix-ui/react-tooltip'
-
-interface TooltipTriggerProps {
-  children: React.ReactNode
-  className?: string
-  asChild?: boolean
-}
-
-export const TooltipTrigger = ({
-  children,
-  className,
-  asChild,
-}: TooltipTriggerProps) => {
-  return (
-    <TooltipPrimitive.Trigger className={className} asChild={asChild}>
-      {children}
-    </TooltipPrimitive.Trigger>
-  )
-}
